Added lead-through script, modified XACRO files, adjusted RAPID for lead-through
Lead-through files allows for the arm to be moved to a location and the postion can then be stored by publishing a set of commands. XACRO files were modified to fix the issue of not having a fixed joint. YuMi and workspace is now attached to a link called Workspace. Changed RAPID files in order to allow only the state servers to be run in order to allow for the arms to be put into lead-through mode for the lead-through script.